+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog for `cuddle`
 
+## 1.1.0.0
+
+* Remove `CTreeRoot'`
+* Changed the type in `CTreeRoot` to a map of resolved `CTree`s
+* Changed the type of the first argument for `generateCBORTerm` and 
+  `generateCBORTerm'` to `CTreeRoot`
+* Removed all exports in `Codec.CBOR.Cuddle.CBOR.Validator` except for 
+  `validateCBOR` and `validateCBOR'`
+
 ## 1.0.0.0
 
 * First official release to Hackage